services/users/requirements.txt
click==6.7
Flask==0.12.2
Flask-Script==2.0.5
greenlet==0.4.12
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
msgpack-python==0.5.1
neovim==0.2.0
Werkzeug==0.14.1
Flask-SQLAlchemy==2.3.2
psycopg2==2.7.3.1
Flask-Testing==0.6.2
gunicorn==19.7.1
coverage==4.4.1
flake8==3.4.1
autopep8==1.3.3
flask-cors==3.0.3
flask-migrate==2.1.1
